moved VegTypeInfo to its own file
