+++ veg2hab/definitietabel.py
@@ -13,7 +13,7 @@ from veg2hab.habitat import HabitatVoorstel
 from veg2hab.io.common import Interface
 from veg2hab.mozaiek import MozaiekRegel, StandaardMozaiekregel
 from veg2hab.vegetatietypen import SBB, VvN
-from veg2hab.vegkartering import VegTypeInfo
+from veg2hab.vegtypeinfo import VegTypeInfo
 
 _LOGGER = logging.getLogger(__name__)
